# Always use analyzer transformations