--- a/analyzer.ml
+++ b/analyzer.ml
@@ -289,7 +289,7 @@ module TexprFilter = struct
 				]) e.etype e.epos
 			in
 			loop e
-		| TWhile(e1,e2,flag) when not (is_true_expr e1) ->
+		| TWhile(e1,e2,flag) when not (flag = NormalWhile && is_true_expr e1) ->
 			let p = e.epos in
 			let e_break = mk TBreak t_dynamic p in
 			let e_not = mk (TUnop(Not,Prefix,Codegen.mk_parent e1)) e1.etype e1.epos in