--- a/tests/unit/compile-each.hxml
+++ b/tests/unit/compile-each.hxml
@@ -5,5 +5,5 @@
 -resource res1.txt@re/s?!%[]))("'1.txt
 -resource res2.bin@re/s?!%[]))("'1.bin
 -dce full
--D analyzer
+#-D analyzer
 -D analyzer-code-motion