This pull request changes the filters so that the analyzer transformations (but not optimizations) are always run. This does a lot of code-cleanup for eliminating temporary variables. Here are some encouraging stats for unit.js:

Compilation  | Size   | $this | Time
------------ | ------ | ----- | ------
Old default  | 1690kb | 798   | 4.065s
Old analyzer | 1670kb | 7     | 6.523s
New default  | 1648kb | 10    | 3.942s
New analyzer | 1632kb | 10    | 4.370s

This loses the JS inlining restriction and closes #1827 among other things!