@delahee 2015-12-29
If it is what I think it is, Congrats for this important step! Thanks!

2015-12-29 11:57 GMT+01:00 Simon Krajewski <notifications@github.com>:

> This pull request changes the filters so that the analyzer transformations
> (but not optimizations) are always run. This does a lot of code-cleanup for
> eliminating temporary variables. Here are some encouraging stats for
> unit.js:
> Compilation Size $this Time Old default 1690kb 798 4.065s Old analyzer
> 1670kb 7 6.523s New default 1648kb 10 3.942s New analyzer 1632kb 10 4.370s
>
> This loses the JS inlining restriction and closes #1827
> <https://github.com/HaxeFoundation/haxe/issues/1827> among other things!
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/HaxeFoundation/haxe/pull/4746
> Commit Summary
>
>    - run analyzer transformation by default
>    - don't forget about do...while(true)
>    - turn do...while(true) to while(true) first
>    - lose JS inlining restriction
>
> File Changes
>
>    - *M* analyzer.ml
>    <https://github.com/HaxeFoundation/haxe/pull/4746/files#diff-0> (3)
>    - *M* filters.ml
>    <https://github.com/HaxeFoundation/haxe/pull/4746/files#diff-1> (123)
>    - *M* optimizer.ml
>    <https://github.com/HaxeFoundation/haxe/pull/4746/files#diff-2> (4)
>    - *M* tests/unit/compile-each.hxml
>    <https://github.com/HaxeFoundation/haxe/pull/4746/files#diff-3> (2)
>
> Patch Links:
>
>    - https://github.com/HaxeFoundation/haxe/pull/4746.patch
>    - https://github.com/HaxeFoundation/haxe/pull/4746.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/HaxeFoundation/haxe/pull/4746>.
>



-- 
David Elahee