@ncannasse 2015-12-30
Before:
```
Total time : 28.948s
------------------------------------
  filters : 3.531s, 12%
  generate swf : 3.473s, 12%
  haxelib : 0.165s, 1%
  macro execution : 16.286s, 56%
  other : 0.650s, 2%
  parsing : 1.929s, 7%
  typing : 2.270s, 8%
  write swf : 0.644s, 2%
```

After
```
Total time : 31.893s
------------------------------------
  analyzer-filter-apply : 0.125s, 0%
  analyzer-filter-unapply : 0.922s, 3%
  analyzer-from-texpr : 1.097s, 3%
  analyzer-to-texpr : 1.130s, 4%
  filters : 3.547s, 11%
  generate swf : 2.915s, 9%
  haxelib : 0.161s, 1%
  macro execution : 16.383s, 51%
  other : 0.691s, 2%
  parsing : 1.895s, 6%
  typing : 2.393s, 8%
  write swf : 0.633s, 2%
```