@ncannasse 2015-12-30
For information, here's the minimal time (time spent just for recompiling without changing anything by using compiler cache). This is usually the time that matters the most to me on a daily basis, since we are doing like 2-3 full compilations but hundreds of re-compilations.

```
Total time : 5.942s
------------------------------------
  filters : 2.017s, 34%
  generate swf : 2.829s, 48%
  haxelib : 0.164s, 3%
  macro execution : 0.001s, 0%
  module cache check : 0.350s, 6%
  other : 0.002s, 0%
  typing : 0.000s, 0%
  write swf : 0.579s, 10%
```