@Simn 2015-12-30
That's roughly what I expected. Can you check if Evoland survives -D analyzer now?