@ncannasse 2015-12-31
Ok, this can be reproduced by compiling + running on Flash the `samples/skin` example. It will fail at runtime in Manager.fillParams which is one of the few areas in heaps that uses some Dynamic things. The JS/WebGL version runs correctly.