@ncannasse 2015-12-31
Also, you might want to look at the unreachable code warnings. 
They come from the following:

```
static inline var BUFFERING = false;
public inline function hasBuffering() {
	return BUFFERING;
}
```

If I had `BUFFERING = #if flash true #else false #endif` instead, I'm quite sure I would get the warning as well. This might be considered a false positive.