@Simn 2015-12-31
How do I test the Flash version of skin? Trying to run it I get

```
VerifyError: Error #1014: Class flash.display3D.textures::RectangleTexture could not be found.

	at h3d.impl::MemoryManager/allocTexture()
	at h3d.mat::Texture/alloc()
	at MethodInfo-1459()
	at hxd.res::Image/loadTexture()
	at hxd.res::Image/toTexture()
	at hxd.res::Any/toTexture()
	at h3d.prim::ModelCache/loadTexture()
	at MethodInfo-831()
	at hxd.fmt.hmd::Library/makeMaterial()
	at hxd.fmt.hmd::Library/makeObject()
	at h3d.prim::ModelCache/loadModel()
	at Main/init()
	at MethodInfo-29()
	at hxd::App/loadAssets()
	at hxd::App/setup()
	at h3d::Engine/onCreate()
	at h3d.impl::Stage3dDriver/onCreate()
```