@Simn 2015-12-31
samples/skin works now. The problem came from some monomorph being bound to h3d.mat.Texture during propagation. I don't really understand why that would cause any issues but I've changed it to not bind monomorphs.

Regarding the unreachable code: It's not exactly a false positive because that code is indeed unreachable on some targets. I think/hope this is a rare case, so maybe having a metadata to disable the warning for specific fields would be the best approach. I don't see what else I could do.