@Simn 2015-12-31
Looks like there has been some regression on hxcpp as well:

```
./src/hxd/snd/_NativeChannel/ChannelMapper.cpp(72) : error C2664: 'void hx::ArrayBase::setUnmanagedData(void *,int)' : cannot convert argument 1 from 'const unsigned char *' to 'void *'
        Conversion loses qualifiers
```

The relevant code:

```haxe
	override function onSample( buf : cpp.Pointer<cpp.Float32>, len : Int ) {
		var data : haxe.io.BytesData = [];
		cpp.NativeArray.setUnmanagedData(data, buf.reinterpret(), len<<2);
		@:privateAccess native.onSample(haxe.io.Float32Array.fromBytes(haxe.io.Bytes.ofData(data)));
	}
```

C++ now:

```c++
::cpp::Pointer< unsigned char > inData = ::cpp::Pointer< unsigned char >(buf->reinterpret());		
 const unsigned char *  tmp = inData->get_raw();
data->setUnmanagedData(tmp,(int(len) << int((int)2))); // line 72
```

C++ before:

```c++
pp::Pointer< unsigned char > inData = ::cpp::Pointer< unsigned char >(buf->reinterpret());
data->setUnmanagedData(inData->get_raw(),(int(len) << int((int)2)));
```