@ncannasse 2015-12-31
Ok, got it. Add the following code in Skin init()

```
			var outline = new h3d.shader.Outline();
			outline.size = 0.;
			outline.distance = 0.022; // not more (cut hero head behind vertical walls)
			for( m in obj.getMaterials() ) {
				var p = new h3d.mat.Pass("outline", m.mainPass);
				p.depth(false, Greater);
				p.addShader(outline);
				p.addShader(new h3d.shader.ZCut(0.45, 0.9));
				m.addPass(p);
			}
```