@Simn 2015-12-31
That gives me this even without -D analyzer:

```
Error(h3d.shader.BaseMesh.__init__5 needed by h3d.shader.ZCut.vertex is unreachable)@null
	at hxsl::Error$/t()
	at hxsl::Linker/error()
	at hxsl::Linker/link()
	at hxsl::Cache/compileRuntimeShader()
	at hxsl::Cache/link()
	at h3d.shader::Manager/compileShaders()
	at h3d.pass::Default/setupShaders()
	at h3d.pass::Default/draw()
	at h3d.scene::Renderer/render()
	at h3d.scene::Renderer/process()
	at h3d.scene::Scene/render()
	at h3d::Engine/render()
	at hxd::App/mainLoop()
	at MethodInfo-29()
	at hxd::App/loadAssets()
	at hxd::App/setup()
	at h3d::Engine/onCreate()
	at h3d.impl::Stage3dDriver/onCreate()
```