@ncannasse 2015-12-31
Uhm, right sorry about it, will look into it might be a regression on my side