@ncannasse 2015-12-31
Ok, Evoland2 seems to run, I didn't do the whole game to check there's no regression, but at least most of the base code seems to run well. Congrats!